main: merge "arguments.py" into "main.py"
"arguments.py" contains "parse_args()", which is the command line
parsing for the "main.py". It makes no sense, that this would ever be
useful to another executable or used somewhere else. This can also be
seen, because the "arguments" module is also only imported by "main.py"
and nowhere else.

This split is unnecessary, and it only makes the code harder to follow,
because you have to switch between files.

Merge them.

The resulting "main.py" is still small enough, to not warrant this
split.
